Fixed manifest creation. Fixed 64-bit warnings.
git-svn-id: svn://svn.code.sf.net/p/shareaza/code/trunk@6801 3231ed4b-2112-0410-84e0-8c3944a40730

+++ torrentaid/Buffer.cpp
@@ -130,7 +130,7 @@ void CBuffer::Clear()
 void CBuffer::Print(LPCSTR pszText)
 {
 	ASSERT( pszText != NULL );
-	Add( (void*)pszText, strlen( pszText ) );
+	Add( (void*)pszText, (DWORD)strlen( pszText ) );
 }
 
 DWORD CBuffer::Append(CBuffer* pBuffer, DWORD nLength)
